i18n: Only use system language as default if in the set { de, en, es, ro }
Closes #1551.

User feedback raised a good point -- defaulting the app to the system
language may be very undesirable for UX as many of the translations we
have are incomplete to the point of being extremely off-putting. A
recent example was Thai which contained nonsense text.

In order to maximize perceived app quality, we just default to English
if the user's system language isn't in the "acceptable for new installs"
set which currently is: Spanish, German, English and Romanian.

Note this change only affects "new installs" that don't yet have a
language that is set.  Users can always set the language manually in the
preferences and their choice will be respected always.